[Rename] 코드 아키텍처 변경
