+++ app/resetpassword/complete/page.tsx
@@ -2,8 +2,8 @@
 
 import { createBrowserSupabaseClient } from 'utils/supabase/client';
 import Head from 'next/head';
-import AuthBackgroundCards from 'components/auth/auth-background-cards';
-import LogoImage from 'components/auth/logo-image';
+import AuthBackgroundCards from 'components/auth/common/auth-background-cards';
+import LogoImage from 'components/auth/common/logo-image';
 
 export default function ResetPasswordCompletePage() {
   const supabase = createBrowserSupabaseClient();
