+++ components/auth/common/logo-image.tsx
