+++ components/layouts/sidebar/footer/index.tsx
@@ -1,12 +1,12 @@
 'use client';
 
 import LogoutButton from './logout-button';
-import Profile from './profile';
+import ProfileBox from './profile-box';
 
 export default function Footer() {
   return (
     <div className={`flex flex-col gap-10 items-center`}>
-      <Profile />
+      <ProfileBox />
       <LogoutButton />
     </div>
   );
