+++ components/layouts/sidebar/sub-sidebar/collected-badge.tsx
