+++ components/layouts/sidebar/sub-sidebar/collected-cafe-detail.tsx
